pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-lab"
version = "0.1.0"
description = "Local research and execution environment for systematic trading."
requires-python = ">=3.11"
dependencies = [
  "anthropic",
  "google-genai",
  "matplotlib",
  "numpy",
  "openai",
  "pandas",
  "plotly",
  "pyarrow",
  "pytest",
  "pytest-cov",
  "python-dotenv",
  "requests",
  "streamlit",
  "ta",
  "trading-ig",
  "yfinance",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src/trading_lab --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source = ["src/trading_lab"]
omit = ["*/tests/*", "*/__init__.py"]
